+++ src/main/java/de/mrapp/tries/PatriciaTrie.java
@@ -28,12 +28,11 @@ import java.util.Map;
 import java.util.NoSuchElementException;
 
 /**
- * A sorted trie, which stores the successor of nodes in sorted lists. In contrast to a {@link
- * SortedListTrie}, the edges between nodes do not always correspond to a single element of a
- * sequence. Instead, subsequent nodes that only have a single successor are merged to a single node
- * to reduce space complexity. This requires to reorganize the tree structure when inserting new
- * elements. Consequently, this trie should be preferred over the a {@link SortedListTrie}, if new
- * elements are only inserted sporadically and minimizing memory consumption is important.
+ * A sorted trie, which stores the successor of nodes in sorted lists. In contrast to a {@link SortedListTrie}, the
+ * edges between nodes do not always correspond to a single element of a sequence. Instead, subsequent nodes that only
+ * have a single successor are merged to a single node to reduce space complexity. This requires to reorganize the tree
+ * structure when inserting new elements. Consequently, this trie should be preferred over the a {@link SortedListTrie},
+ * if new elements are only inserted sporadically and minimizing memory consumption is important.
  *
  * @param <SequenceType> The type of the sequences, which are used as the trie's keys
  * @param <ValueType>    The type of the values, which are stored by the trie
@@ -49,15 +48,14 @@ public class PatriciaTrie<SequenceType extends Sequence, ValueType>
     private static final long serialVersionUID = 3229102065205655196L;
 
     /**
-     * Traverses the trie in order to return the node, which corresponds to a specific sequence. If
-     * the sequence does not correspond to a node, the next higher node with the same prefix is
-     * returned.
+     * Traverses the trie in order to return the node, which corresponds to a specific sequence. If the sequence does
+     * not correspond to a node, the next higher node with the same prefix is returned.
      *
-     * @param sequence The sequence, the node, which should be returned, corresponds to, as an
-     *                 instance of the generic type {@link SequenceType} or null
-     * @return A pair, which contains the node, which corresponds to the given sequence,
-     * respectively the next higher node, as well as the key, the returned node corresponds to, as
-     * an instance of the class {@link Pair} or null, if the given sequence does not match any node
+     * @param sequence The sequence, the node, which should be returned, corresponds to, as an instance of the generic
+     *                 type {@link SequenceType} or null
+     * @return A pair, which contains the node, which corresponds to the given sequence, respectively the next higher
+     * node, as well as the key, the returned node corresponds to, as an instance of the class {@link Pair} or null, if
+     * the given sequence does not match any node
      */
     @Nullable
     private Pair<Node<SequenceType, ValueType>, SequenceType> getSubTrieRootNode(
@@ -80,15 +78,14 @@ public class PatriciaTrie<SequenceType extends Sequence, ValueType>
                 } else {
                     currentNode = pair.first;
                     suffix = pair.second;
-                    matchedPrefix = suffix == null || suffix.isEmpty() ? sequence : SequenceUtil
-                            .subsequence(sequence, 0, sequence.length() - suffix.length());
+                    matchedPrefix = suffix == null || suffix.isEmpty() ? sequence :
+                            SequenceUtil.subsequence(sequence, 0, sequence.length() - suffix.length());
                 }
             }
 
             if (sequence != null && !sequence.isEmpty() && matchedPrefix != null &&
                     matchedPrefix.length() < sequence.length()) {
-                SequenceType unmatchedSuffix =
-                        SequenceUtil.subsequence(sequence, matchedPrefix.length());
+                SequenceType unmatchedSuffix = SequenceUtil.subsequence(sequence, matchedPrefix.length());
                 int index = currentNode.indexOfFirstElement(unmatchedSuffix);
 
                 if (index != -1) {
@@ -107,11 +104,10 @@ public class PatriciaTrie<SequenceType extends Sequence, ValueType>
     /**
      * Creates a new Patricia trie.
      *
-     * @param rootNode   The root node of the trie as an instance of the type {@link Node} or null,
-     *                   if the trie should be empty
-     * @param comparator The comparator, which should be used to compare keys with each other, as an
-     *                   instance of the type {@link Comparator} or null, if the natural ordering of
-     *                   the keys should be used
+     * @param rootNode   The root node of the trie as an instance of the type {@link Node} or null, if the trie should
+     *                   be empty
+     * @param comparator The comparator, which should be used to compare keys with each other, as an instance of the
+     *                   type {@link Comparator} or null, if the natural ordering of the keys should be used
      */
     protected PatriciaTrie(@Nullable final Node<SequenceType, ValueType> rootNode,
             @Nullable final Comparator<? super SequenceType> comparator) {
@@ -126,11 +122,11 @@ public class PatriciaTrie<SequenceType extends Sequence, ValueType>
     }
 
     /**
-     * Creates a new Patricia trie, which contains all key-value pairs that are contained by a map.
-     * For comparing keys with each other, the natural ordering of the keys is used.
+     * Creates a new Patricia trie, which contains all key-value pairs that are contained by a map. For comparing keys
+     * with each other, the natural ordering of the keys is used.
      *
-     * @param map The map, which contains the key-value pairs that should be added to the trie, as
-     *            an instance of the type {@link Map}. The map may not be null
+     * @param map The map, which contains the key-value pairs that should be added to the trie, as an instance of the
+     *            type {@link Map}. The map may not be null
      */
     public PatriciaTrie(@NotNull final Map<SequenceType, ValueType> map) {
         super(null, map);
@@ -139,9 +135,8 @@ public class PatriciaTrie<SequenceType extends Sequence, ValueType>
     /**
      * Creates a new, empty Patricia trie.
      *
-     * @param comparator The comparator, which should be used to compare keys with each other, as an
-     *                   instance of the type {@link Comparator} or null, if the natural ordering of
-     *                   the keys should be used
+     * @param comparator The comparator, which should be used to compare keys with each other, as an instance of the
+     *                   type {@link Comparator} or null, if the natural ordering of the keys should be used
      */
     public PatriciaTrie(@Nullable final Comparator<? super SequenceType> comparator) {
         super(comparator);
@@ -150,11 +145,10 @@ public class PatriciaTrie<SequenceType extends Sequence, ValueType>
     /**
      * Creates a new Patrica trie, which contains all key-value pairs that are contained by a map.
      *
-     * @param comparator The comparator, which should be used to compare keys with each other, as an
-     *                   instance of the type {@link Comparator} or null, if the natural ordering of
-     *                   the keys should be used
-     * @param map        The map, which contains the key-value pairs that should be added to the
-     *                   trie, as an instance of the type {@link Map}. The map may not be null
+     * @param comparator The comparator, which should be used to compare keys with each other, as an instance of the
+     *                   type {@link Comparator} or null, if the natural ordering of the keys should be used
+     * @param map        The map, which contains the key-value pairs that should be added to the trie, as an instance of
+     *                   the type {@link Map}. The map may not be null
      */
     public PatriciaTrie(@Nullable final Comparator<? super SequenceType> comparator,
             @NotNull final Map<SequenceType, ValueType> map) {
@@ -175,16 +169,18 @@ public class PatriciaTrie<SequenceType extends Sequence, ValueType>
 
     @NotNull
     @Override
-    public final SortedTrie<SequenceType, ValueType> subTrie(
-            @Nullable final SequenceType sequence) {
+    public final SortedTrie<SequenceType, ValueType> subTrie(@Nullable final SequenceType sequence) {
         Pair<Node<SequenceType, ValueType>, SequenceType> pair = getSubTrieRootNode(sequence);
 
         if (pair != null) {
             Node<SequenceType, ValueType> node = pair.first;
+            boolean hasSuccessors = node.hasSuccessors();
+            boolean includeNodeValue =
+                    (sequence != null && pair.second != null && pair.second.length() > sequence.length());
 
-            if (node.hasSuccessors()) {
+            if (hasSuccessors || includeNodeValue) {
                 Node<SequenceType, ValueType> rootNode =
-                        structure.getSubTrie(pair.second, createRootNode(), node);
+                        structure.getSubTrie(pair.second, createRootNode(), node, includeNodeValue && !hasSuccessors);
                 return new PatriciaTrie<>(rootNode, comparator);
             } else {
                 return new PatriciaTrie<>(null, comparator);