+++ RELEASE_NOTES.md
@@ -1,5 +1,11 @@
 # Tries - RELEASE NOTES
 
+## Version 1.0.1 (Feb. 2nd 2018)
+
+A bugfix release, which fixes the following issues:
+
+- Fixed an issue, which may cause the sub tries of a `PatriciaTrie` or `PatriciaStringTrie` to lack some values.
+
 ## Version 1.0.0 (Jan. 30th 2018)
 
 The first stable release of the library, which provides the following features: